小文件合并重构,修复文件数量，以及文件大小过大的情况下，整个程序无法跑起来的情况......
